Move tests to folders
