+++ src/Tests/Tests.Dybal.Utils.Guards/ArgumentGuard/UnitTestsBase.cs
@@ -1,6 +1,6 @@
 ﻿using Xunit;
 
-namespace Tests.Dybal.Utils.Guards;
+namespace Tests.Dybal.Utils.Guards.ArgumentGuard;
 
 [Trait("Category", "Unit")]
 public class UnitTestsBase
